Move type Page from redux feature into typings/data/import.ts
Update dependencies
