+++ docs/src/modules/components/md/MdDocs.tsx
@@ -2,7 +2,7 @@ import { createStyles, makeStyles, Theme } from '@material-ui/core';
 import { useHoux } from 'houx';
 import React from 'react';
 
-import { Page } from '../../redux/features/navigation/type';
+import { Page } from '../../../../../src/typings/data/import';
 import { RootState } from '../../redux/reducers';
 import AppContent from '../AppContent';
 import AppFrame from '../AppFrame';
