+++ package.json
@@ -41,6 +41,7 @@
     "clsx": "^1.0.4",
     "compose-middleware": "^5.0.1",
     "compression": "^1.7.4",
+    "deepmerge": "^4.0.0",
     "envalid": "^5.0.0",
     "gray-matter": "^4.0.2",
     "houx": "^1.1.2",
@@ -49,7 +50,7 @@
     "lodash": "^4.17.15",
     "mdast-util-to-string": "^1.0.6",
     "mdast-util-toc": "^4.2.0",
-    "next": "^9.0.7-canary.2",
+    "next": "^9.0.7-canary.4",
     "next-i18next-serverless": "^1.1.33",
     "prism-react-renderer": "^0.1.7",
     "react": "^16.9.0",
@@ -96,7 +97,7 @@
     "nodemon": "^1.19.2",
     "prettier": "^1.18.2",
     "raw-loader": "^3.1.0",
-    "ts-loader": "^6.1.0",
+    "ts-loader": "^6.1.2",
     "ts-node": "^8.4.1",
     "typescript": "^3.6.3",
     "webpack": "^4.40.2",
