+++ src/typings/data/import.ts
@@ -1,5 +1,15 @@
 import { GridSize } from '@material-ui/core/Grid';
 
+export interface Page {
+  pathname: string;
+  title?: string;
+  children?: Array<Page>;
+  displayNav?: boolean;
+  subheader?: string;
+  icon?: string;
+  highlight?: boolean;
+}
+
 export interface Allocation {
   step: number;
   size: GridSize;
